fix(exg-gwf*): check for equal idomain  (#2149)
Check that models connected to a flow model by an exchange have an idomain identical to the flow model's idomain. This is an easy error to make, and (up to now) hard to detect.

I only added a test for PRT for now. The logic is identical in GWT and GWE.

#2144 facilitates adding the same check for flow model interface. That will be done in a separate PR since it deserves a separate release note.