docs(prt-fmi): clarify only required when post-processing (#2151)
The first paragraph in the fmi sections of mf6io could be read to indicate that the fmi file is always required. Clarify that the file is only required when the corresponding model is run as a post-processor to the flow model. Also update a prt test to make sure the head/budget file paths provided via fmi are ignored when the simulation uses an exchange