fix(utl-hpc.dfn): fix flopy subpackage metadata (#2182)
I think the final term here should be hpc_data, to follow the convention used for other subpackages and produce the generated files as they currently appear in flopy — at least assuming the same code is to handle all subpackages. Seems like there is inconsistent handling currently, depending whether they are attached to simulation or to a model or package.